refactor(ai): OpenAI 통신 로직 및 GPT 관련 책임 분리
- OpenAiClient 클래스 생성하여 OpenAI API 통신 로직 인프라 레이어로 분리
- GptPromptBuilder 생성하여 프롬프트 생성 책임 분리
- GptResponseParser 생성하여 GPT 응답 파싱 책임 분리
- AICourseGenerationService를 오케스트레이션 역할로 슬림화
- 도메인 서비스와 인프라 레이어 간 의존성 명확히 구분
